Moved src and header files to appropriate subdirectories
    Moved to Objects: 'snake' and 'border'
    Moved to Helper: 'renderer' and 'controller'
    Not moved: 'game' and 'main'
