+++ src/Helper/controller.h
@@ -1,7 +1,7 @@
 #ifndef CONTROLLER_H
 #define CONTROLLER_H
 
-#include "snake.h"
+#include "../Objects/snake.h"
 
 class Controller {
  public:
